Added partial support for file-level headings via org-roam
Since file-level headings are an org-roam-ism that technically aren't proper org syntax, it makes sense to use functions from org-roam which are mean specifically to deal with them. In many places, an incompatible function has been replaced with some variation of (or (incompatible-function) (ignore-errors (org-roam-compatible-function).
A roadblock to progress here is org-map-entries, which is used multiple times and has no ready compatible alternative.